CREATE and TEST database layer

+++ src/main/java/com/virtadity/manease/domain/model/Product.java
@@ -1,11 +1,12 @@
 package com.virtadity.manease.domain.model;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record Product (
     UUID productId,
     String name,
-    Double weight,
+    BigDecimal weight,
     UUID producerId,
     UUID productTypeId
 ) {}
